Create separate package for appenders
